# openapi.yaml
openapi: 3.1.0
info:
  title: BuyWhere Catalog API
  description: |
    Agent-native product catalog API for AI agent commerce.
    Query millions of products across Southeast Asia.
  version: 1.0.0
  contact:
    name: BuyWhere Engineering
    url: https://buywhere.ai
  license:
    name: Proprietary
servers:
  - url: https://api.buywhere.ai
    description: Production
  - url: https://staging-api.buywhere.ai
    description: Staging
security:
  - BearerAuth: []
...
paths:
...
  /v1/search:
    get:
      tags:
        - search
      summary: Full-text search across product catalog
      description: |
        Full-text search using PostgreSQL ts_rank with English language ranking.
        Supports filtering by category, price range, platform, and availability.
        Results cached for 10 minutes.
        Rate limit: 1000 requests/minute per API key.
      operationId: searchProducts
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          description: Full-text search query
          required: false
          schema:
            type: string
            example: iphone 15 pro
        - name: category
          in: query
          description: Filter by category (case-insensitive partial match)
          required: false
          schema:
            type: string
            example: Electronics
        - name: min_price
          in: query
          description: Minimum price filter
          required: false
          schema:
            type: number
            format: decimal
            minimum: 0
            example: 100
        - name: max_price
          in: query
          description: Maximum price filter
          required: false
          schema:
            type: number
            format: decimal
            minimum: 0
            example: 999
        - name: platform
          in: query
          description: Filter by platform/source
          required: false
          schema:
            type: string
            enum:
              - lazada_sg
              - shopee_sg
              - carousell
              - qoo10
            example: shopee_sg
        - name: in_stock
          in: query
          description: Filter by availability
          required: false
          schema:
            type: boolean
            example: true
        - name: limit
          in: query
          description: Number of results per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: offset
          in: query
          description: Pagination offset
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
      responses:
        '200':
          description: Product search results
          headers:
            X-RateLimit-Limit:
              description: Rate limit per minute
              schema:
                type: integer
                example: 1000
            X-RateLimit-Remaining:
              description: Remaining requests in current window
              schema:
                type: integer
                example: 999
            X-RateLimit-Reset:
              description: Unix timestamp when rate limit resets
              schema:
                type: integer
                example: 1709568060
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
              example:
                total: 1523
                limit: 20
                offset: 0
                has_more: true
                items:
                  - id: 12345
                    sku: SHOPEE-SG-IPHONE15PRO-256-BLK
                    source: shopee_sg
                    merchant_id: shop_abc123
                    name: Apple iPhone 15 Pro 256GB Black
                    description: Latest Apple iPhone 15 Pro with A17 Pro chip
                    price: 1499
                    currency: SGD
                    buy_url: https://shopee.sg/iphone-15-pro
                    affiliate_url: https://affiliate.buywhere.ai/track/123
                    image_url: https://cf.shopee.sg/file/abc123
                    category: Electronics
                    category_path:
                      - Electronics
                      - Phones
                      - Smartphones
                    is_available: true
                    metadata:
                      brand: Apple
                      model: iPhone 15 Pro
                    updated_at: '2026-04-03T10:30:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/products:
    get:
      tags:
        - products
      summary: Search products
      description: |
        Full-text search across product catalog with optional filters.
        Uses PostgreSQL full-text search with English language ranking.
        Rate limit: 1000 requests/minute per API key.
      operationId: searchProductsLegacy
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          description: Full-text search query
          required: false
          schema:
            type: string
            example: iphone 15 pro
        - name: category
          in: query
          description: Filter by category name (case-insensitive partial match)
          required: false
          schema:
            type: string
            example: Electronics
        - name: min_price
          in: query
          description: Minimum price filter
          required: false
          schema:
            type: number
            format: decimal
            minimum: 0
            example: 100
        - name: max_price
          in: query
          description: Maximum price filter
          required: false
          schema:
            type: number
            format: decimal
            minimum: 0
            example: 999
        - name: source
          in: query
          description: Filter by source platform
          required: false
          schema:
            type: string
            enum:
              - lazada_sg
              - shopee_sg
              - carousell
              - qoo10
            example: shopee_sg
        - name: limit
          in: query
          description: Number of results per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: offset
          in: query
          description: Pagination offset
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
      responses:
        '200':
          description: Product search results
          headers:
            X-RateLimit-Limit:
              description: Rate limit per minute
              schema:
                type: integer
                example: 1000
            X-RateLimit-Remaining:
              description: Remaining requests in current window
              schema:
                type: integer
                example: 999
            X-RateLimit-Reset:
              description: Unix timestamp when rate limit resets
              schema:
                type: integer
                example: 1709568060
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
              example:
                total: 1523
                limit: 20
                offset: 0
                items:
                  - id: 12345
                    sku: SHOPEE-SG-IPHONE15PRO-256-BLK
                    source: shopee_sg
                    merchant_id: shop_abc123
                    name: Apple iPhone 15 Pro 256GB Black
                    description: Latest Apple iPhone 15 Pro with A17 Pro chip
                    price: 1499
                    currency: SGD
                    buy_url: https://shopee.sg/iphone-15-pro
                    affiliate_url: https://affiliate.buywhere.ai/track/123
                    image_url: https://cf.shopee.sg/file/abc123
                    category: Electronics
                    category_path:
                      - Electronics
                      - Phones
                      - Smartphones
                    is_available: true
                    metadata:
                      brand: Apple
                      model: iPhone 15 Pro
                    updated_at: '2026-04-03T10:30:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/categories:
    get:
      tags:
        - categories
      summary: List available categories
      description: |
        Returns all product categories with active product counts.
        Sorted by product count descending.
        Rate limit: 1000 requests/minute per API key.
      operationId: listCategories
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Category list with product counts
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
                example: 1000
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: 999
            X-RateLimit-Reset:
              schema:
                type: integer
                example: 1709568060
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
              example:
                categories:
                  - name: Electronics
                    count: 45230
                    children: []
                  - name: Fashion
                    count: 32100
                    children: []
                  - name: Home & Living
                    count: 21500
                    children: []
                total: 48
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or API key
      description: |
        Authentication via API key or JWT token.
        - For direct API keys: use the key prefixed with `bw_` as-is
        - For JWT tokens: obtain via OAuth flow, contains `key_id` claim
        All API requests must include this header.
  schemas:
...
    ProductListResponse:
      type: object
      required:
        - total
        - limit
        - offset
        - items
        - has_more
      properties:
        total:
          type: integer
          description: Total number of matching products
          example: 1523
        limit:
          type: integer
          description: Results per page
          example: 20
        offset:
          type: integer
          description: Pagination offset
          example: 0
        has_more:
          type: boolean
          description: Whether there are more results available
          example: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
...
  responses:
    Unauthorized:
      description: Invalid or missing API key
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Invalid or revoked API key
          example:
            detail: Invalid or revoked API key
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
          example: 60
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: 'Rate limit exceeded: 1000/minute'
          example:
            detail: 'Rate limit exceeded: 1000/minute'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Internal server error
          example:
            detail: Internal server error